renamed statusxxx to the status name
