+++ routes/api/auth.js
@@ -3,14 +3,14 @@ const gitServices = rootRequire("services/git");
 const mongodbService = rootRequire("services/mongodb");
 const validation = rootRequire("utils/validation");
 const { generateToken } = rootRequire("utils/jwt");
-const errors = rootRequire("consts/errors");
+const { clientError } = rootRequire("consts/errors");
 const { convertToErrorModel } = rootRequire("utils/error");
 
 const userSignin = async (req, res) => {
   try {
     const authProvider = req.body.authProvider;
     if (!validation.authProvider(authProvider))
-      throw errors.status400.INVALID_AUTH_PROVIDER;
+      throw clientError.INVALID_AUTH_PROVIDER;
     const gitService = gitServices[authProvider];
     const code = req.body.code;
     const token = await gitService.getToken(code);
@@ -29,7 +29,7 @@ const userSignin = async (req, res) => {
     return res.json({ token: jwt });
   } catch (err) {
     err = convertToErrorModel(err);
-    console.log(err);
+    console.log(`❗ ${err}`);
     return res.status(err.status).json(err);
   }
 };
