+++ services/git/github/auth.js
@@ -1,10 +1,10 @@
 const axios = require("axios");
 const { GITHUB_CLIENT_ID, GITHUB_CLIENT_SECRET } = rootRequire("config");
 const validation = rootRequire("utils/validation");
-const errors = rootRequire("consts/errors");
+const { clientError, serverError } = rootRequire("consts/errors");
 
 const getToken = async code => {
-  if (!validation.githubCode(code)) throw errors.status400.INVALID_GITHUB_CODE;
+  if (!validation.githubCode(code)) throw clientError.INVALID_GITHUB_CODE;
   try {
     const data = {
       client_id: GITHUB_CLIENT_ID,
@@ -20,7 +20,7 @@ const getToken = async code => {
     return res.data.access_token;
   } catch (err) {
     console.log(`❗ ${err}`);
-    throw errors.status500.GITHUB_ERROR;
+    throw serverError.GITHUB_ERROR;
   }
 };
