+++ services/mongodb/user.js
@@ -1,9 +1,9 @@
 const User = rootRequire("models/mongodb/User");
 const validation = rootRequire("utils/validation");
-const errors = rootRequire("consts/errors");
+const { clientError, serverError } = rootRequire("consts/errors");
 
 const getUser = async (provider, id, createIfDosntExist = false) => {
-  if (!validation.githubUserId(id)) throw errors.status400.INVALID_USER_ID;
+  if (!validation.githubUserId(id)) throw clientError.INVALID_USER_ID;
   try {
     const authProvider = `authProviders.${provider}`;
     const user = await User.findOne({ [authProvider]: id });
@@ -11,7 +11,7 @@ const getUser = async (provider, id, createIfDosntExist = false) => {
     return user;
   } catch (err) {
     console.log(`❗ ${err}`);
-    throw errors.status500.MONGO_DOCUMENT_NOT_FOUND;
+    throw serverError.MONGO_DOCUMENT_NOT_FOUND;
   }
 };
 
@@ -23,7 +23,7 @@ const createUser = async (authProvider, id) => {
     return user;
   } catch (err) {
     console.log(`❗ ${err}`);
-    throw errors.status500.MONGO_ERROR;
+    throw serverError.MONGO_ERROR;
   }
 };
