+++ utils/error.js
@@ -1,8 +1,8 @@
-const errors = require("../consts/errors");
+const { serverError } = require("../consts/errors");
 const { ErrorModel } = rootRequire("models/error");
 
 const convertToErrorModel = error => {
-  const defaultError = errors.status500.SERVER_ERROR;
+  const defaultError = serverError.SERVER_ERROR;
   return new ErrorModel(
     error.status || defaultError.status,
     error.code || defaultError.code,
